- Extended Euclidean Algorithm implemented and working for the cases where the School Division works.
 - School division bug found. When attempting to divide 1 by 1, the rest
is 1, and the quotant is left untouched (undefined).

 - Bug found and solved in the cleanZeros() method. Previously it was
allowed to attempt to access the element with index 0 for data vectors
of size 0 (with no elements), hence, an exception was thrown. It's
corrected now.

(It's late now, but if the lingering bug with the division isn't found,
I'll get myself busy with it and devise my own school algorithm)